Express internally new vs legacy approach to keys' locking via an interface
A keysLockSolution let's execute or try the authorization policy facet corresponding
to the strategy and indexes in use for sealed keys locking.

Move sealed keys' locking code to keyslock.go.
